Split into relevant files and PEP8 compliance
